23-08-2023 Final Code

+++ BanasVehicleTrackDbClasses/BanasAuditorVisitGatepassRetrieve_Result.cs
@@ -28,17 +28,17 @@ namespace BanasVehicleTrackDbClasses
         public string Driver { get; set; }
         public string VehicleCode { get; set; }
         public string VehicleRegNumber { get; set; }
-        public Nullable<decimal> StartOdometer { get; set; }
+        public Nullable<int> StartOdometer { get; set; }
         public string CreateUser { get; set; }
         public Nullable<System.DateTime> CreateDate { get; set; }
         public int GatePassStatus { get; set; }
-        public Nullable<decimal> CloseOdometer { get; set; }
+        public Nullable<int> CloseOdometer { get; set; }
         public string CloseRemark { get; set; }
         public string CloseDateTime { get; set; }
         public string CloseDate { get; set; }
         public string CloseUser { get; set; }
-        public Nullable<decimal> Netkm { get; set; }
-        public Nullable<decimal> Difference { get; set; }
+        public Nullable<int> Netkm { get; set; }
+        public Nullable<int> Difference { get; set; }
         public Nullable<int> DepartmentId { get; set; }
         public Nullable<int> VehicleId { get; set; }
         public int VisitPurposeId { get; set; }
@@ -47,14 +47,14 @@ namespace BanasVehicleTrackDbClasses
         public string DepartureSecuritySignature { get; set; }
         public int DepartureSecurityId { get; set; }
         public string DepartureSecurityRemark { get; set; }
-        public Nullable<decimal> StartKm { get; set; }
+        public Nullable<int> StartKm { get; set; }
         public string DepartureDateTime { get; set; }
         public int DepartureVerifyStatus { get; set; }
         public string ArrivalSecurityName { get; set; }
         public string ArrivalSecuritySignature { get; set; }
         public int ArrivalSecurityId { get; set; }
         public string ArrivalSecurityRemark { get; set; }
-        public Nullable<decimal> EndKm { get; set; }
+        public Nullable<int> EndKm { get; set; }
         public string ArrivalDateTime { get; set; }
         public int ArrivalVerifyStatus { get; set; }
         public string EmployeeSignature { get; set; }
